refactor(network): 将MessageOperationType从枚举类型改为byte类型
简化消息操作类型的存储和传输，使用byte类型替代枚举类型以提高性能和减少序列化开销

+++ GameFrameX.Hotfix/StartUp/AppStartUpHotfixGameByMain.cs
@@ -105,7 +105,7 @@ internal partial class AppStartUpHotfixGame
             }
 
             var actorId = netWorkChannel.GetData<long>(GlobalConst.ActorIdKey);
-            if (outerNetworkMessage.Header.OperationType == MessageOperationType.HeartBeat)
+            if (outerNetworkMessage.Header.OperationType == (byte)MessageOperationType.HeartBeat)
             {
                 if (Setting.IsDebug && Setting.IsDebugReceive && Setting.IsDebugReceiveHeartBeat)
                 {
